+++ src/lab6/P3.java
@@ -2,9 +2,9 @@ package lab6;
 
 public class P3 {
  
-	static float multi(float a, float b)
+	static double multi(double a, double b)
 	{	
-		float  multi = a * b;             ////Autor Domino kielbasa i wino :)
+		double  multi = a * b;             ////Autor Domino kielbasa i wino :)
 		return multi;
 	}
